working on searching pins
