+++ src/javascripts/components/Boards.js
@@ -2,7 +2,6 @@ const showBoards = (arr) => {
   document.querySelector('#content-container').innerHTML = '';
   document.querySelector('#form-container').innerHTML = '';
   document.querySelector('#add-button').innerHTML = '<div class="dropdown"><button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">Create</button><div class="dropdown-menu" aria-labelledby="dropdownMenuButton"><a class="dropdown-item" href="#" id="create-board-btn">Board</a><a class="dropdown-item" href="#" id="create-pin-btn">Pin</a></div></div>';
-  // WILL NEED TO ADD BUTTON
   arr.forEach((item) => {
     document.querySelector('#content-container').innerHTML += `<div class="card" style="width: 18rem;">
     <img src="${item.image}" class="card-img-top" alt="${item.title}">
